Overlays

+++ snippets/stream/Example_bad_foreach.java
@@ -1,12 +1,12 @@
-public static Map<User, Integer> findFriendsAverageAgeOfUsersWithNameHeinz(List<User> users) {
-    Map<User, Integer> results = new HashMap<>();
+public static Map<User, Double> findFriendsAverageAgeOfUsersWithNameHeinz(List<User> users) {
+    Map<User, Double> results = new HashMap<>();
     users.forEach(user -> {
             if (user.getName().equals("Heinz")) {
-                int sumAge = 0;
+                double sumAge = 0;
                 user.getFriends().forEach(friend -> {
                         sumAge += friend.getAge();
                     });
-                int averageAge = sumAge / user.getFriends().size();
+                double averageAge = sumAge / user.getFriends().size();
                 results.put(user, averageAge);
             }
         });
